refactor: rename trace2 to trace and deprecate old trace view
- Rename /pages/trace to /pages/trace-old
- Rename /pages/project/[projectId]/traces to /pages/project/[projectId]/traces-old
- Rename /pages/project/[projectId]/traces2 to /pages/project/[projectId]/traces
- Update navigation paths in TracePage to use /traces instead of /traces2
- Remove duplicate /traces2/[traceId] entry from publishable paths

This makes the new trace view the default at /traces URL while keeping
the old trace view accessible at /traces-old for backwards compatibility.
