+++ web/src/components/layouts/app-layout/utils/pathClassification.ts
@@ -15,7 +15,6 @@ export const PATH_CONSTANTS = {
   publishable: [
     "/project/[projectId]/sessions/[sessionId]",
     "/project/[projectId]/traces/[traceId]",
-    "/project/[projectId]/traces2/[traceId]",
     "/auth/reset-password",
   ] as const,
 };
