+++ web/src/pages/project/[projectId]/traces/[traceId].tsx
@@ -1,4 +1,4 @@
-import { TracePage } from "@/src/components/trace/TracePage";
+import { TracePage } from "@/src/components/trace2/TracePage";
 import { useRouter } from "next/router";
 
 export default function Trace() {
